+++ src/Objects/Events.php
@@ -6,69 +6,49 @@ use andrefelipe\Orchestrate\Application;
 
 class Events extends AbstractList
 {
-
-    /**
-     * @var string
-     */
-    protected $key;
+    use KeyTrait, TypeTrait;
 
 
 
-
-    public function __construct(Application $application, $collection, $key=null)
+    public function __construct(Application $application, $collection, $key=null, $type=null)
     {
         parent::__construct($application, $collection);
         $this->key = $key;
+        $this->type = $type;
     }
-
-    /**
-     * @return string
-     */
-    public function getKey()
-    {
-        return $this->key;
-    }
-
-    /**
-     * @param string $key
-     */
-    public function setKey($key)
-    {
-        $this->key = $key;
-    }
-
 
 
 
     /**
      * @param int $limit
-     * @param int $offset
-     * @param boolean $values
-     * @return Refs self
+     * @param array $range
+     * @return Events self
      */
-    public function listEvents($type, $limit=10, $startEvent='', $afterEvent='', $beforeEvent='', $endEvent='')
+    public function listEvents($limit=10, array $range=null)
     {
         // required values
         $this->noCollectionException();
         $this->noKeyException();
-        
+        $this->noTypeException();
+
         // define request options
-        $path = $this->collection.'/'.$this->key.'/events/'.$type.'/';
+        $path = $this->collection.'/'.$this->key.'/events/'.$this->type.'/';
         
         $parameters = ['limit' => $limit];
 
-        if ($startEvent)
-            $parameters['startEvent'] = $startEvent;
+        if ($range) {
+            if (isset($range['start']))
+                $parameters['startEvent'] = $range['start'];
 
-        if ($afterEvent)
-            $parameters['afterEvent'] = $afterEvent;
+            if (isset($range['after']))
+                $parameters['afterEvent'] = $range['after'];
 
-        if ($beforeEvent)
-            $parameters['beforeEvent'] = $beforeEvent;
+            if (isset($range['before']))
+                $parameters['beforeEvent'] = $range['before'];
 
-        if ($endEvent)
-            $parameters['endEvent'] = $endEvent;       
-
+            if (isset($range['end']))
+                $parameters['endEvent'] = $range['end'];
+        }
 
         // request
         $this->request('GET', $path, ['query' => $parameters]);