+++ src/Objects/TypeTrait.php
@@ -1,0 +1,40 @@
+<?php
+namespace andrefelipe\Orchestrate\Objects;
+
+/**
+ * Trait that implements the Type methods
+ */
+trait TypeTrait
+{
+
+    /**
+     * @var string
+     */
+    protected $type = null;
+    
+
+    /**
+     * @return string
+     */
+    public function getType()
+    {
+        return $this->type;
+    }
+
+    /**
+     * @param string $type
+     */
+    public function setType($type)
+    {
+        $this->type = $type;
+    }
+
+
+    protected function noTypeException()
+    {
+        if (!$this->type) {
+            throw new \BadMethodCallException('There is no type set yet. Please do so through setType() method.');
+        }
+    }
+    
+}