+++ CHANGELOG.md
@@ -1,5 +1,11 @@
 # CHANGELOG
 
+## 0.2.0-alpha - 2015-01-04
+
+- Implemented List, Refs and Events
+- Added pagination methods for all List operations (search, events, refs)
+
+
 ## 0.1.0-alpha - 2014-12-31
 
 - Initial version